Change namespace to CharBrowser
