+++ include/CharBrowser/Repositories/ItemRepository.php
@@ -1,8 +1,24 @@
 <?php
 
-namespace Magelo\Repositories;
+/***************************************************************************
+ *
+ *   This program is free software; you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation; either version 2 of the License, or
+ *   (at your option) any later version.
+ *
+ *   Portions of this program are derived from publicly licensed software
+ *   projects including, but not limited to phpBB, Magelo Clone,
+ *   EQEmulator, EQEditor, and Allakhazam Clone.
+ *
+ *                                  Author:
+ *                           Maudigan(Airwalking)
+ *
+ ***************************************************************************/
 
-use Magelo\Stores\ItemStore;
+namespace CharBrowser\Repositories;
+
+use CharBrowser\Stores\ItemStore;
 
 class ItemRepository
 {
