+++ include/CharBrowser/Stores/ItemStore.php
@@ -1,0 +1,29 @@
+<?php
+
+/***************************************************************************
+ *
+ *   This program is free software; you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation; either version 2 of the License, or
+ *   (at your option) any later version.
+ *
+ *   Portions of this program are derived from publicly licensed software
+ *   projects including, but not limited to phpBB, Magelo Clone,
+ *   EQEmulator, EQEditor, and Allakhazam Clone.
+ *
+ *                                  Author:
+ *                           Maudigan(Airwalking)
+ *
+ ***************************************************************************/
+
+namespace CharBrowser\Stores;
+
+class ItemStore
+{
+    use BaseStoreTrait;
+
+    /**
+     * @var string
+     */
+    private static $namespace = 'items';
+}
