--- include/Magelo/Stores/CharacterAlternateAbilityStore.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Magelo\Stores;
-
-class CharacterAlternateAbilityStore
-{
-    use BaseStoreTrait;
-
-    /**
-     * @var string
-     */
-    private static $namespace = 'character_alternate_ability';
-}
