--- include/Magelo/Stores/CharacterFactionValueStore.php
@@ -1,15 +1,0 @@
-<?php
-
-
-namespace Magelo\Stores;
-
-
-class CharacterFactionValueStore
-{
-    use BaseStoreTrait;
-
-    /**
-     * @var string
-     */
-    private static $namespace = 'character_faction_value';
-}
