--- include/Magelo/Stores/ItemStore.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Magelo\Stores;
-
-class ItemStore
-{
-    use BaseStoreTrait;
-
-    /**
-     * @var string
-     */
-    private static $namespace = 'items';
-}
