--- include/Magelo/Stores/SpellStore.php
@@ -1,15 +1,0 @@
-<?php
-
-
-namespace Magelo\Stores;
-
-
-class SpellStore
-{
-    use BaseStoreTrait;
-
-    /**
-     * @var string
-     */
-    private static $namespace = 'spells';
-}
